changed Price to decimal and implemented `ListAllMovies()` query.

+++ MoviesMenuSql/Models/Movie.cs
@@ -1,6 +1,6 @@
 ﻿namespace MoviesMenuSql.Models;
 
-public class Movie(int? Id, string? Title, string? Director, string? Genre, int? ReleaseYear, double? Price)
+public class Movie(int? Id, string? Title, string? Director, string? Genre, int? ReleaseYear, decimal? Price)
 {
 
     public int? Id { get; set; } = Id;
@@ -8,6 +8,6 @@ public class Movie(int? Id, string? Title, string? Director, string? Genre, int?
     public string? Director { get; set; } = Director;
     public int? ReleaseYear { get; set; } = ReleaseYear;
     public string? Genre { get; set; } = Genre;
-    public double? Price { get; set; } = Price;
+    public decimal? Price { get; set; } = Price;
 
 }
